Campaign perms (#8)
* Check campaign permissions with superuser override

* break off can as its own module

* replace with can

+++ controllers/campaign_controller.js
@@ -1,6 +1,6 @@
 const Campaign = require('../models/campaign')
-const client = require('../config/contentful').default
 const map = require('lodash/map')
+const can = require('../middleware/can')
 
 exports.index = async (req, res, next) => {
   const query = { $or: [
@@ -24,20 +24,40 @@ exports.store = async (req, res, next) => {
 }
 
 exports.show = async (req, res, next) => {
-  const campaign = await Campaign.findById(req.params.id)
-    .populate('empire.grandAdmiral')
-    .populate('empire.players')
-    .populate('rebels.grandAdmiral')
-    .populate('rebels.players')
-  return res.json(campaign)
+  try {
+    const campaign = await Campaign.findById(req.params.id)
+      .populate('empire.grandAdmiral')
+      .populate('empire.players')
+      .populate('rebels.grandAdmiral')
+      .populate('rebels.players')
+    if (!campaign) return res.status(404)
+    return res.json(campaign)
+  } catch(err) {
+    return res.status(500).json(err)
+  }
 }
 
 exports.update = async (req, res, next) => {
-  const campaign = await Campaign.findByIdAndUpdate(req.params.id, req.body, { new: true })
-  return res.json(campaign)
+  try {
+    const campaign = await Campaign.findById(req.params.id)
+    if (!campaign) return res.status(404)
+    if (!can('update:campaign', req.user, campaign.user)) return res.status(403)
+    const updatedCampaign = await Campaign.findByIdAndUpdate(req.params.id, req.body, { new: true })
+    return res.json(updatedCampaign)
+  } catch (err) {
+    return res.status(500).json(err)
+  }
 }
 
 exports.destroy = async (req, res, next) => {
-  await Campaign.findByIdAndDelete(req.params.id)
-  return res.status(204).json(null)
+  try {
+    const campaign = await Campaign.findById(req.params.id)
+    if (!campaign) return res.status(404)
+    if (!can('delete:campaign', req.user, campaign.user)) return res.status(403)
+    await Campaign.findByIdAndDelete(req.params.id)
+    return res.status(204).json(null)
+  } catch(err) {
+    console.log(err)
+    return res.status(500).json(err)
+  }
 }
